Force download of file regardless of type
This change removes the code that treats PDFs as a special case.
Instead, PDF files are treated exactly the same as any other file and
are downloaded directly rather than attempting to utility browser-based
PDF viewers.

This avoids the problem of popup blockers getting in the way of PDF
downloads when using non-Safari browsers in Bitwarden Send.